requirements.txt
requests
pillow
discord_webhook
dotenv